+++ src/com/zhgl/run/service/EventViolationService.java
@@ -8,10 +8,10 @@ import com.zhgl.util.dao.DAO;
 
 public interface EventViolationService extends DAO<EventViolation> {
 
-	public HashMap<String, String> getTypeCount(int sid);
+	public HashMap<String, String> getTypeCount(long sid);
 
 	public HashMap<String, String> getTypeCount(Date beginDate, Date endDate,
-			int sid);
+			long sid);
 
 	/**
 	 * 统计指定塔机今日违章数量
@@ -19,10 +19,10 @@ public interface EventViolationService extends DAO<EventViolation> {
 	 * @param tcsid
 	 * @return
 	 */
-	public long countToday(int sid);
+	public long countToday(long sid);
 
 	/**
 	 * 查出在此时间范围内,此塔机违章的数量
 	 */
-	public int getCount(Date beginDate, Date endDate, int sid);
+	public int getCount(Date beginDate, Date endDate, long sid);
 }
